Refactor Konoha town placement helpers
